+++ docs/en/changes/changes.md
@@ -17,9 +17,10 @@
 * Add support for OpenSearch/ElasticSearch client certificate authentication.
 * Fix BanyanDB logs paging query.
 * Replace BanyanDB Java client with native implementation.
-* Remove `bydb.dependencies.properties` and set the compatible BanyanDB API version number in `${SW_STORAGE_BANYANDB_COMPATIBLE_SERVER_API_VERSIONS}`
+* Remove `bydb.dependencies.properties` and set the compatible BanyanDB API version number in `${SW_STORAGE_BANYANDB_COMPATIBLE_SERVER_API_VERSIONS}`.
 * Fix trace profiling query time range condition.
 * Fix BanyanDB time range overflow in profile thread snapshot query.
+* `BrowserErrorLog`, OAP Server generated UUID to replace the original client side ID, because Browser scripts can't guarantee generated IDs are globally unique.
 * MQE: fix multiple labeled metric query and ensure no results are returned if no label value combinations match.
 
 #### UI